+++ CHANGELOG.md
@@ -14,6 +14,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Adjusted shader database to define output layers recursively to more accurately represent in game shaders.
 * Improved accuracy, code size, and generation time for xc3_wgpu shader generation.
 
+### Removed
+* Removed dependency list from `xc3_model::shader_database` and associated methods. Use the first layer if present for detecting dependencies instead.
+
 ## 0.19.0 - 2025-04-03
 ### Added
 * Added support for wimdo models from Xenoblade X DE.